Don't move DeviceInfo
